# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movieflix API",
        "description": "API responsible to manage the movies of Movieflix",
        "version": "1.0.0"
    },
    "paths": {
        "/movies": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "To get all the movies",
                "responses": {
                    "200": {
                        "description": "Return a movies list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "put": {
                "tags": [
                    "Movie"
                ],
                "summary": "To update a movie data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID of the movie that will be updated",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Movie updated sucessfully"
                    },
                    "404": {
                        "description": "Movie doesn't found"
                    },
                    "500": {
                        "description": "It was not possible to update the movie register"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movie"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID of the movie that will be deleted",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie deleted sucessfully"
                    },
                    "404": {
                        "description": "Movie doesn't found"
                    },
                    "500": {
                        "description": "It was not possible to delete the movie"
                    }
                }
            }
        },
        "/movies/{genreName}": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "To filter movies by gender",
                "parameters": [
                    {
                        "name": "genreName",
                        "in": "path",
                        "required": true,
                        "description": "Movie gender name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "It was not possible to show the movies list filtered by gender"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Movie ID"
                    },
                    "title": {
                        "type": "string",
                        "description": "Movie title"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "Movie genre"
                    },
                    "genres": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "Genre ID"
                            },
                            "name": {
                                "type": "string",
                                "description": "Genre name"
                            }
                        }
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "Movie language"
                    },
                    "languages": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "Language ID"
                            },
                            "name": {
                                "type": "string",
                                "description": "Language name"
                            }
                        }
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "How many Oscars the movie received"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Movie release date"
                    }
                }
            },
            "MovieInput": {
                "type": "object",
                "required": [
                    "title",
                    "genre_id",
                    "language_id",
                    "oscar_count",
                    "release_date"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Movie title"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "Movie ID"
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "Movie language"
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "How many Oscars the movie received"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Movie release date"
                    }
                }
            }
        }
    }
}